Made The Entire Game
I did some stuff because Huxs is bad at giving directions
