+++ Abilities/MagmaBlast.java
