+++ Attributes/DoubleJumps/DirectDoubleJump.java
