+++ Attributes/DoubleJumps/DoubleJump.java
