+++ Attributes/DoubleJumps/GenericDoubleJump.java
