+++ Attributes/ExpCharge.java
