+++ Attributes/Regeneration.java
